decouple SetLookup stage and data structure
