+++ lucille-core/src/main/java/com/kmwllc/lucille/stage/LookupCollection.java
@@ -1,0 +1,10 @@
+package com.kmwllc.lucille.stage;
+
+public interface LookupCollection<T> {
+
+  boolean add(T value);
+
+  boolean contains(T value);
+
+  boolean remove(T value);
+}
